Decouple worker accounting from scheduler
This patch is based on two patches found in PREEMPT_RT kernels

1. sched: Distangle worker accounting from rqlock
2. workqueue: Prevent deadlock/stall on RT

the first patch is to move the worker accounting for cpu bound workers out
of the core scheduler code and moving it into the workqueue code. Keep track
of the sleeping state in the worker itself and call the notifier before
entering the core scheduler.

The second patch is protect / serializes the list of idle workers which can be
woken up, by placing locks around the idle list functions.

Signed-off-by: ninez <johnstonljordan@gmail.com>