scheduler rework/updates
this all comes from my local 4.9-eas kernel. There is still a lot
of code to go through; adjust and/or upgrade... However, this patch covers
some immediate spots that needed to be fixed up, re-ordered and/or made easier
to read; when doing comparisons (diff)...

Also, updated schedutil, while I was at it.

Signed-off-by: ninez <johnstonljordan@gmail.com>